+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+### Changed
+### Deprecated 
+### Removed
+### Fixed
+### Security
+
+
+## [2.3.0]
+### Added
 - [issue/126](https://github.com/podaac/l2ss-py/issues/126): Added flexibility to variable subsetting
 for variables to not have leading slash in the front
 - [issue/136](https://github.com/podaac/l2ss-py/issues/136): Added type annotations throughout the package code